+++ lib/bolt/consumer/ready.ex
@@ -101,11 +101,11 @@ defmodule Bolt.Consumer.Ready do
     },
 
     ## Server filter management
-    "filter" => %{
-      "add" => Cogs.Filter.Add,
-      "show" => Cogs.Filter.Show,
-      "remove" => Cogs.Filter.Remove
-    },
+    # "filter" => %{
+    # "add" => Cogs.Filter.Add,
+    # "show" => Cogs.Filter.Show,
+    # "remove" => Cogs.Filter.Remove
+    # },
 
     ## Rule verification
     "accept" => Cogs.Accept,